+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to the
 [Haskell Package Versioning Policy](https://pvp.haskell.org/).
 
-## UNRELEASED
+## 0.3.0
 
 * Change default language to `Haskell2010`.
 * Remove deprecated `PEM` instance of type class `NormalForm` and the direct